This is a rework of the files generated by SUCASA.  Before, SUCASA generated three files:
  app_milp.h/app_milp.cpp   Define the derived MILP classes
  app_extras.cpp            Used to define the methods specialized
                                by the user

Now, SUCASA generates five files:

  app_sucasa.h/app_sucasa.cpp   Define the derived MILP classes
  app_milp.h/app_milp.cpp       Define a MILP from the MILP derived in
                                    the *_sucasa.* files.
  app_extras.cpp                Used to define the methods specialized
                                    by the user

The result, is that SUCASA now generates two derived solvers.  BUT,
the one that is exposed to the user can be customized without
impacting the integration of the Info classes.  Further, this segregation
simplifies the class definitions that a user looks at; for example, there
are no ugly ifdefs, and no references to the Info data.