Reworked the way that 'empty' objectives and constraints are dealt with.  By default, objectives and constraints with expr/body values of None are provided, which facilitates some aspects of preprocessing.  But this triggered exceptions, which are now disabled.
Reworked how the '_equality' constraint attribute is managed.
This is now in the ConstraintData class, which required various
changes.

Changed the pprint output for equality constraints to make it clear 
that both the upper and lower bounds are equal for these constraints.

Fixed bug in the definition of parameter names.  When using a rule to
define parameter values, the parameter name now reflex the index
values for the parameter.