Adding support for specifing options when launching solvers.   For example:
  results = self.pico.solve(currdir+"bell3a.mps", options="maxCPUMinutes=0.1")

I've added tests, but these are not applied automatically.  It wasn't clear how
to easily test that the command-line changed as expected, especially given the
diverse array of solvers and how they are configured with options.

Another fix was the change that when pyomo is specified with the '-l' option, then
all files are kept;  that is, '-l' triggers the '-k' options.  Without this logic,
'-l' generated errors without the '-k' flag.

Also, the '--solver-options' option was added to pyomo, which takes a string.  Note that this option can be applied multiple times, and that it concatenates the option values.