Fixed an issue in the WW PH extension involving variable slamming. During initialization, variable values are assigned the associated stage for purposes of slamming to force convergence.
When picking a variable to slam, the loop looks for the "_stage" attribute - which is only defined for variables appearing in all but the last stage. 

However, the specification for variable slamming in wwph.suffixes input files may necessarily inadvertently (for reasons of simplicity) specify variable values in the final stage, e.g., foo[*,*].

I now check for the existence of the "_stage" attribute before blinding accessing it.