I have significantly changed - and improved - the handling of variable lower and upper bounds.
In the old version of pyomo, if parameters were supplied as bounds, the lb/ub were silently ignored when writing LP/NL/etc files. 

In the new version of pyomo, these parameters are stored and handled to avoid this issue. In particular, the "lb" and "ub" attributes of a variable are either (1) None, (2) a NumericConstant, or (3) a _ParamBase. The latter two objects share the common () operator. Thus, to access the lb and ub attributes of a variable, one should either test for None, or use lb() or ub() to extract the value. 

To enforce the ability to use () on any non-None bound, I have provided two "setter" methods for variables: setlb(value) and setub(value), where value is an instance of one of the objects indicated above. We normally would trap the set with the __setattr__ functionality Python provides, but the overhead of this function is too steep for general use. Thus, there is no explicit enforcement of the setlb/setub method route to set lower and upper bounds - if you don't do this, errors will arise when the values are accessed (e.g., when writing an LP/NL file - specifically errors related to the () operator not being defined for a certain type).

Expressions are easy to add now, but a few points have to be discussed before proceeding.