Bug fix:  If Python version is < 2.6.4, then checkout Coopr Forum with http.  Otherwise, use https.
There's a deeper issue here.  A bug fix in 2.6.3 allowed https URLs to 
be opened through a proxy.  This is exactly what we need at Sandia, so 
unless we have Python 2.6.4 installed everywhere, it doesn't make sense to
use https for Coopr Forum.

I've asked to have 2.6.4 installed on tofu, since that will allow us to develop
Coopr Forum stuff there.

NOTE:  other users may run into this problem when using coopr_install.  I'll wait
for that to happen, but when/if it does, we'll probably need to rework the
coopr_install configuration to use pypi rather than https URLs for release 
branches.