Changed an error message to be more informative.
Thought: the reason the original message wasn't very useful was
because pyomo can't show me the line in my model file that the error
is coming from; at best, pyomo shows the stack trace going through all
the different pyomo functions. This traceback is good for developers,
but not terribly useful for an end user debugging a model. Is there a
way to get pyomo to reference model file lines on errors?