Changes to coopr.pyomo to support SOS constraints.
util.py was altered to attach the 'has_capability' function from the chosen
solver to the model, which allows the writer to know whether the solver
supports SOS constraints or not.

cpxlp.py was altered to output SOS constraints in CPLEX format when asked.

constraint.py was altered to add a new base class for constraint objects, ConstraintBase, as well as a new derived constraint, SOSConstraint.