Commit new magic number to keep development Logger silent.
As noted in the comment around the offending code, the magic number
business is temporary until we complete the Logger update.  There's
some more thought that needs to go into the Logging system, but in
the mean time, let's not frustrate users who aren't uber developers
but use the trunk version.