Adds Piecewise variables.
Piecewise objects are variables that are constrained to have a value
determined by a piecewise linear function. If the piecewise linear
function is concave or convex, the variable is constrained by a number
of lienar constraints. If it is nonconvex, then it is comstrained
using SOS2 sets.

TODO: Test this class rigorously.