On the road to Coopr with Python 2.4
Another NFC:  Another couple of naive and not-so-naive tricks to
bring to 80 character width code.  In the process removed a couple
of isinstance calls (constraint.py) and remembered (haduh!) some
XOR logic (cpxlp).  Was too late when I saw the deprecated around
the function in question, but oh well; iteration is good, right?