Refs #3881: "CPlex LP line length enforcement"
Formally enforce CPlex's arbitrary 510 char line length limit CPlex.
This still fails to test or properly mangle 255-character
identifiers, but that might be enforceable at a different layer.
Haven't completely through that one out yet.

Of note, this commit also assumes that if people are going to want
to read the LP files, it tries its best to stick to 80 characters
wide (good for screen reading and printing to paper).