Fixes an apparent bug.
Before, arithmetic operations on expressions were defined to operate
in-place, which causes issues when manipulating expressions. For example,

> foo = model.X + model.Y
> for x in xrange(0,10):
>     print foo + model.X

would in fact manipulate the expression foo. As this is believed to be
undesirable (and was motivated by a specific use case, namely model
transformations) it has been changed to return a clone of the
expression in question.

One test case failed as a side effect of this change, and a correction
was made to resolve this. This correction was also inside expr.py. The
problem was that the ._nargs attribute of _SumExpression objects was
not copied on .clone() calls. This has been changed. NOTE: it is not
known whether a similar fix must be made to product expressions. No
test cases are failing now, but this must be resolved.