Fixes bug in Pyomo reading unbounded/infeasible solution files from GLPK.
Bug fixed as a side effect while formatting code to fix the bug,
indicating that it was likely caused by a poorly-indented `raise`
statement, or some other source of an Exception that was resolved via
new indentation.

It is worthwhile here to note that our solution may still read as
'infeasible' when the problem is in fact known to be unbounded; this
is due to GLPK's presolver, and short of turning off GLPK's presolver
this is beyond Pyomo's abilities to fix.