Finalizing the design of immutable parameters!  This design follows CEP 4. All model components are stored as attributes of model.components, and immutable parameters are stored as float and dictionary attributes of the model object.
This should make references to parameter values more intuitive.
Specifically, you can directly reference singleton parameter values,
rather than use the value() function.  Additionally, these are constant values,
and thus they can be used in boolean expressions without generating
Pyomo expression objects.

HOWEVER, immutable parameter values do not have the rich representation/context
that is provided by the Pyomo Param class.  For example, suppose that we have:

  Z =[0,1]
  model.A = Param(Z, default=1.0)

Then to query the index set for A you would need to do:

  model.components.A.keys()

rather than

  model.A.keys()

Both may be syntactically correct, but model.A is a dictionary whose
keys may be a subset of all valid keys for model.components.A.

The model attributes for immutable parameters are either (a) floats
(for singleton parameters) or (b) dictionaries (for indexed parameters).
Furthermore, the dictionary object used for indexed parameters has been
enhanced to support default values in a sparse manner.


Other changes:

- The Block class was restructured to contain the 'components' attribute,
which itself is a new class that manages the indices for blocks.
Some of the functionality in Block was migrated into BlockComponents
(which is used to define the 'components' attribute).

- The semantics of the summation() function has been restricted somewhat.
If the 'index' option is not specified, then this function now validates
that the last argument is a variable component.  This is necessary to
ensure that this argument contains a complete index set.

- Many tests were updated due to these changes.  However, these were
almost exclusively the tests for the parameter component itself.