Adding the __nonzero__ method to NumericValue and Expression objects. This allows logic like:
 if model.p > 1:
    pass

to work correctly.  The "model.p > 1" string creates an expression, whose
value is evaluated with the __nonzero__ method when the condition is
evaluated.

Additionally, error checking was added when the value of a NumericValue
object is requested but the value is None.  This alerts the user when
they are accessing data objects that haven't been initialized.

Some tests were added ... and some tests were fixed since the error checking
highlighted that they were not working correctly.