Fixing label issues associated with recent GDP failures, or more correctly, partially fixes.
When _add_component is called, the component name is updated. However, the label was not. I added "sync_label" methods to the Var/ConstraintData/ObjectiveData classes, which are then called from within _add_component. This automatically keeps the (default) label synced with the name.

This fix - the presence of a label with a name, independent of preprocessing - also exposed test issues related to uninstantiated models. In particular, the abstract models used in those tests are being translated into concrete models via the pyomo script, and all is working fine because these models don't rely on external data for construction. This is obviously not the desired behavior, as the model shouldn't be constructed. Those tests have been temporarily suppressed, as it exposed a number of unexpected issues - to be addressed in the next commit.