Removing Expression.simplify()  - all simplification functions occur within generate_expression()  - expression trees are actually slightly "more simplified" than before  - the new simplification routine is measurably faster than the old    recursive simplify() methods.
It is an open question as to if we want to preserve expression
simplification.  Currently, generating simplified expression trees is
slightly slower (5% - 15%) than generating raw, unsimplified expression
trees.  However, the simplified trees are smaller and more readable.  As
is indicated in #4176, it is still debatable if we should perform basic
expression simplification at all.