First stab at speeding up canonical expressions   - check expr.polynomial_degree() before collecting the expression   - adding a specialized collector for linear expressions   - pause the Python GC
These changes should be 100% backward compatible for polynomial
expressions.  Non-polynomial expressions are now returned intact in the
`[None]` group (that is, the polynomial terms are not extracted) and all
variables in the expression are added to the `[-1]` group.  While this
behavior is somewhat different from the original canonical expression,
no one should notice as there are no NL writers that use canonical
expressions.