The cost of solution variables being stored as a SolutionMap is no longer tenable, particularly for large-scale problems. When solving some big hydro problems from Chile, it was taking over an hour just to create the solution! Similar issues exist on all large-scale problems that I've tested, and show up in our profiles on moderate-scale problems. So it is time to jump ship - once this stabilizes, we will probably need to make the same change for constraints.
The change is conceptually subtle - just replace SolutionMap for a native Python dictionary. However, there is a lot of automatic name-related functionality in SolutionMap that disappears - and other functionality related to __setattr__ and __getattr__ of SolutionMaps is no longer supported. In particular:
1) "soln.variable.x" is now "soln.variable["x"]"
2) "soln.variable.x.value" is now "soln.variable["x"]["Value"]". soln.variable["x"]=3 (for example) is no longer supported - it would nuke the dictionary associated with "x".
The new syntax roughly corresponds to what we in place around 1.5 years ago, prior to the introduction of SolutionMaps.

For anyone creating a Coopr solution (basically, only solver plugin developers), they need to know that "Value" and "Id" are the two attributes they need to define for each variable. And they must follow that specific naming convention, as Python dictionaries of course don't do any type of string lower-casing. Id values are assumed to be 0-based.

These changes are likely to be disruptive to testing, although I presently only see a handful of failures. I'm committing now, to let Hudson help me find the remaining issues. The remaining errors are unlikely to be observed by any current user. 

These changes basically eliminate the run-time associated with solution creation, i.e., dropping hour-plus times to a second or so. 

More commits to follow, for the other Coopr components impacted. I'll be cleaning up remaining test issues over the weekend.