This commit is a significant change to the API of coopr.pyomo classess. The default API for construction rules, validation functions, filter functions, etc has been changed to put the model arguement first.
The previous API placed indices first and then the model.  This has
several problems:

. This is counter-intuitive from a programming standpoint: in all
languages (including Python) you always put the variable-length arguments
at the end of the argument list. That is, we should do "model, I, J, …".

. This is SLOWER. We spend a measurable amount of time setting up the
arguments for rules that place the model last.  The indices are converted
to a list, the model is added to the end, and this list is converted
back to a tuple to pass to the rule.

. This places a restriction on rules: either all invocations of a rule
must use the same number of indices, or the rule must explicitly take
*args and remove model from the end itself. If the model came first,
users could define the rule as: "def _rule(model, *index):"

This commit changes the semantics while still maintaining a degree
of backwards compatibility.  We use try-except blocks to detect when
a rule fails due to an incorrect argument sequence.  In this manner,
Old-style models will often still run (albeit, slightly slower), but
new-style models would be both cleaner and faster.

Currently, a warning log message is generated when an old-style rule
is used.  Later, we can make this an error log (in Coopr 3.1), before removing the
try-except block altogether in future Pyomo releases (in Coopr 4.0).

This commit resolves #4214 and #4163