Reworking the set operators.  These operators now generate a new set object that is (a) virtual but (b) indicates that concrete values are available.  The _SetContainer object was reworked to weaking the dependencies on the 'value' attribute, and the _SetOperator class defines accesses to the set value via an iterator.  This makes some things, like 'len' computations expensive.  However, the sets generated in this fashion can be cached in another set object.  For example:
  model.A = Set(initialize=model.B - model.C)

I've used pympler to confirm that the operator expressions do not generate data
that the old operator expressions generated.  However, in total the new set classes
require more memory.  I suspect that this is due to the fact that _SetOperator and its
subclasses are PCA plugins ... which incurr some fixed overhead.  But I haven't been 
able to figure this out a more concrete explanation after looking at this for a while...