Reworking the implementation of the RangeSet class.  This class is now a virtual concrete class, which means that the set elements are not explicitly constructed.  This offers a space savings, and in most cases a time savings (at least for construction).
However, this changes some fundamental behavior in RangeSet:

1. You cannot add, remove or discard elements of a RangeSet, since it is
virtual.

2. If the filter or validate methods are specified, then the cost
of using RangeSet objects becomes much more expensive (i.e. as
expensive as the old implementation).  Specifically:

  a. All elements need to be generated to determine the set length
  b. Set membership needs to call these methods

If the filter and validate functions are expensive, then the RangeSet
object may be more expensive to use than the old implementation.


Since 99% of all uses of RangeSet look like the following:

  RangeSet(1,10)

this should be a big improvement for the typical user.