Adding support for FuncDesigner functions for tanh, asinh, acosh and atanh. Adding support for Pyomo intrinsic functions for pow, abs, floor and ceil. Now, Pyomo and FuncDesigner are roughly sync'd in the functions they support.
Tweaked the way that intrinsic functions are generated.  The test_terms.py file
was failing on the pow() test because I was passing in a number.  This didn't make
sense to me, so I changed how the arguments are processed when creating intrinsic functions.
But, this causes a new test failure, which I need to discuss with John.