Migrating suffix logic from var.py to indexedcomponent.py
This allows all indexed components to manage suffixes.

I change the terminology from 'attribute' to 'suffix'.  This is consistent with
the documentation of other modeling tools (e.g. GAMS, AMPL and AIMMS).

I renamed fixed_value() to is_fixed(), which is consistent with the naming
of similar methods in Pyomo.