Modifications to allow for fixed variable lower and upper bounds. Primary goal is to correct issue with mutable params being illegal.
We still need to confirm that variable bounds are indeed still fixed at solve time, but that is a down-stream / future issue.

Corrected an issue with fixed variables still being written to LP files. Caused no errors, but yielded a bunch of annoying warning messages.

Fixed an issue with the benders example, due to mutable=False default changes.