A fairly major re-structuring of canonical expressions, in order to make a significant impact on its rather large memory footprint.
The biggest change is the introduction of a LinearCanonicalRepn object, which does exactly what you might think - it is a specialized version of the general dictionary, for linear expressions. Doing this alone drops the memory requirements for pmedian8 from 700MB to 600MB (due to the elimination of a dictionary per object). Does require some switch logic in the writers, but those are isolated - mainly because only the LP writer uses canonical expressions. 

Also sets the stage for 200MB of additional reductions, but I won't commit those changes until Jenkins clears up from this commit.

Summary: Python dictionaries are evil! Or, more precisely, if you have a lot of them (or objects with them)...