Culling the 'index' attribute/slot from VarData, ConstraintData, and ObjectiveData classes. These are not used except in non performance-critical cases (e.g., name/label generation), and consume unnecessary memory. The same information is now obtainable via the index() method on ConstraintData instances.
On SparseIndexedComponent and IndexedComponent, changed the name of the index() method to index_set(). The latter is more accurate (it returns the index set of a component, not an individual index). Fortunately, index() is not referenced anywhere in the Pyomo book, and is only used in a handful of places within Coopr. Subsequent commits will address those changes.