A fairly major re-structuring of the core PH code, focusing on a tighter integration of the scenario tree class hierarchy with the scenario instances themselves. In theory, no functional change.
This commit sets the framework for removing of the reference instance, and improvements in the weight and variable average computations in PH (among other things). 

The main thrust of the re-work was to get all references to the reference instance out of the scenario tree, which instead now points to scenario instance _VarValue objects.