Modifications of Var component relating to bulk values store/extract routines.
Modified extract_values method to accept a keyword argument indicating whether to extract values for fixed variables. The ability to do so saves time and network bandwidth in PH solver server contexts.

Modified store_values method to work correctly when a subset of the indices are specified in the input dictionary.