Adding activate/deactivate functionality to _BlockData:   - activate()/deactivate() on indexed Blocks works the same     way as with indexed Constraints. However, calling (de)activate()     on a Block does not call the same on all Block subcomponents     (e.g., Constraints, Objectives, Blocks), so it is up to the writer     or transformation to only iterate (top down) over active blocks.   - *** Modifying the all_blocks() method to, by default, iterate over     only active Blocks/_BlockData. ***   - Removing (de)activate() methods from Piecewise and _PiecewiseData     as these are subclasses of Block and _BlockData, which now support     this functionality.
Note that the .active flag has always existed on Block (but not _BlockData). And Pyomo
was inconsistent about when it was checked (e.g., only the LP writer used it). By adding
this functionality to _BlockData, I am simply trying to enforce consistent behavior between
a non-indexed Block and a particular index of an indexed Block, which are conceptually used
for the same things.