Bug fix when running parallel ph using phboundextension with variable fixing (e.g., wwphextension):   - Now caching the status and value of fixed variables in ph solution cache   - Solution caching no longer occurs by default in order to not slow     down ph when not using phboundextension. This is activated by setting     ph._enable_solution_caching = True in something like a ph_initialization callback
Also fixing an issue when running parallel ph with bundling and the phboundextension. The bundle binding instance objectives
can't be used in this case because no bundle binding instances are created! Therefore, we shall always compute the objective
bound using the individual scenario objective (this theoretically gives the same answer, whether or not we're bundling)