Fixing tickets #4323 and #4482.  We used to support renaming or moving a component simply by assigning it to a new block or by assigning it to a new attribute on the same block (although we tossed a warning).
Attempting this now raises an exception.

Justification for the change in behavior: we implemented the move /
rename by internally performing "block.del_component() +
block.add_component()".  This has the net effect of moving the component
to the END of the construction order -- a very bad thing if it is
something like a Set where other components were counting on the set to
already be constructed (which gave rise to the error reported in #4482)!