Fix for when the --handshake-with-phpyro runph option is used:
The Problem:
The list of action handles returned by the initialize_ph_solver_servers function is passed to the solver manager's wait_all() method at the end of ph.initialize() in order to synchronize the master ph instance with the phsolverservers. However, when the --handshake-with-phpyro option was used, an intermediate call to wait_all() occurred with the action handles used to transmit the master scenario tree ids, causing the queued results associated with the initialize_ph_solver_server action handles to be dropped (during subsequent calls to wait_any()). When this happened, the final call to wait_all with these action handles would hang indefinitely.

The Fix:
Don't call wait_all() inside the transmit_scenario_tree_ids function. Instead, return the action handles created by this function and add them to the list of action handles past to wait_all() at the end of ph.initialize