Changing the semantics of the DataManagerFactory(), and the way that DataManager plugins are registered.
Now, all plugins are created.  None of them are conditionally created.

The DataManagerFactory() functor now returns UnknownDataManager()
if no data manager plugin is known.  If the plugin is known but not
available, then an exception is raised with information about the
python packages that are required to use that data manager.